+++ docs/changelog.rst
@@ -12,6 +12,7 @@ Added
 
 New library methods:
 
+-  :func:`ocdskit.hierarchy.get_base_classes_via_fca`
 -  :func:`ocdskit.normalize.convert_from_oas3`
 -  :func:`ocdskit.normalize.fix_validation_errors`
 -  :func:`ocdskit.normalize.get_base_class_name`